Stop using replaceAll function (way to modern...)
* and move some misplaced functions
